#10 #11 change the structure of factory to simplify and started doing the menu and functionalitis for the backoffice

+++ jobs4u.persistence.impl/src/main/java/domain/AdminEntity.java
@@ -7,7 +7,7 @@ public class AdminEntity {
     @GeneratedValue
     @Id
     @Column(name = "IDADMIN")
-    private int idAdmin;
+    private Long idAdmin;
     @Basic
     @Column(name = "PASSWORD")
     private String password;
@@ -31,7 +31,7 @@ public class AdminEntity {
 
     }
 
-    public int getIdAdmin() {
+    public Long getIdAdmin() {
         return idAdmin;
     }
